move reward to order fulfilled
